chore: MovieInformationResult test commit
